requirements.txt
PyYAML==6.0
schedule==1.1.0
pywin32==223